% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R, R/equal_sample.R
\name{brewer_color_sets}
\alias{brewer_color_sets}
\title{生成颜色系列}
\usage{
brewer_color_sets(n)

brewer_color_sets(n)
}
\arguments{
\item{n}{需要产生的颜色种类数量}
}
\value{
返回指定数量的颜色序列

返回指定数量的颜色序列
}
\description{
RColorBrewer Set1,Set2,Set3合在一起使用，生成包含29种差异较大的颜色的系列

RColorBrewer Set1,Set2,Set3合在一起使用，生成包含29种差异较大的颜色的系列
}
