% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{cluster_marker_preview}
\alias{cluster_marker_preview}
\title{cluster_marker_preview}
\usage{
cluster_marker_preview(
  rawdata,
  groups,
  all_markers,
  cluster_color = dif_seq_rainbow,
  expression_color = rev(brewer.pal(n = 7, name = "RdYlBu")),
  cluster_name = "metacluster",
  output_dir = "cluster_marker_preview",
  major_cond = "Tissue_Type",
  summerise_method = "median",
  Rowv = T,
  Colv = T,
  cluster_id = NULL,
  xlim = NULL,
  free_x = F,
  colored_by = "cluster",
  trans_method = "simpleAsinh"
)
}
\arguments{
\item{groups}{实验组的Metadata信息,其中必需包含列名：”Short_name“}

\item{cluster_color}{tSNE图的亚群的配色方案，默认dif_seq_rainbow}

\item{cluster_name}{用过分析的cluster名称}

\item{output_dir}{输出数据文件夹名称}

\item{major_cond}{选择groups中一个列名做为展现差异的major_cond}

\item{cluster_id}{指定显示的cluster}

\item{xlim}{指定x轴范围，例如：c(0,5)，只在free_x=F的时候有效}

\item{free_x}{True或者FALSE，设置是否每张图X-轴取值范围是否自动，建议设置为F}

\item{trans_method}{数据转化方式，有四种："CytofAsinh"，"simpleAsinh"；"0_to_Max"，所有Marker的信号强度除以最大值，线性转换，通过除以各通道信号的最大值把数值scale到0~1；"Min_to_Max"，线性转换，最小值转换成0，最大值转换成1，最大限度展现population的表达差异}

\item{xdata}{数据框或者矩阵，附带有(meta)cluster表达矩阵}
}
\value{
none
}
\description{
预览各个cluster中marker的表达，以heatmap和直方图的形式展示（密度图）；
}
