% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{cluster_merge}
\alias{cluster_merge}
\title{cluster_merge}
\usage{
cluster_merge(cluster_result, merge_list, cluster_rename = T)
}
\arguments{
\item{cluster_result}{Vector变量，带合并处理cluster}

\item{merge_list}{list变量，指定要合并的cluster名称,例如：list(c(1,2,3),c(6:9)) 将1,2,3合并，将6到9合并；}

\item{cluster_rename}{TRUE/FALSE，决定是否对合并后的cluster重新命名}
}
\description{
手动合并cluster
}
