% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{combined_stat}
\alias{combined_stat}
\title{combined_stat 统计分析函数}
\usage{
combined_stat(cond1_data, cond2_data, alternative = "two.sided",
  mu = 0, stat.paired = FALSE, conf.level = 0.95,
  stat.method = "t-test", set.equal.var = "FALSE",
  p.adjust.method = "BH", silent = F)
}
\arguments{
\item{cond1_data}{一个包含数字的向量（vector），待统计分析数据}

\item{cond2_data}{一个包含数字的向量（vector），待统计分析数据，如进行单组分析，cond2_data=NULL}

\item{alternative}{一个字符串，反应假设检验的种类,有三种取值："two.sided" (默认), "greater" or "less"}

\item{mu}{一个数字，在单样本本统计时表示平均值的真实值，或者如果您正在进行两个样本测试，则表示均值差异（此时一般取值为0）}

\item{stat.paired}{逻辑变量，TRUE或者FALSE，表示是否是成对检验}

\item{conf.level}{显著性水平，默认0.95,（即p<0.05为显著）}

\item{stat.method}{一个字符串，指定统计分析的方法，取值有三种："auto"，“t-test"，“wilcox-test”,当取值为"auto"时，即为测试模式，会对数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。}

\item{set.equal.var}{一个字符串，指定方差齐性，取值有三种："TRUE","FALSE"(默认),"auto",选择"auto"时即可以进行方差齐性测试；}

\item{p.adjust.method}{对p值进行校正的方法："BH"(默认，推荐) "holm", "hochberg", "hommel", "bonferroni","BY","fdr", "none"}

\item{silent}{逻辑变量，是否输出中间结果}
}
\value{
返回一个数据框（data.frame),包含p_value,p_adjust,统计方法(stat.method),方差齐性（set.equal.var）信息
}
\description{
对进行单组或两组数据进行统计分析，并对结果进行adjust。也可以为数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。
}
