% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_cytofkit_functions.R
\name{cytof_addToFCS_modified}
\alias{cytof_addToFCS_modified}
\title{cytof_addToFCS_modified （modified from cytofkit::cytof_addToFCS）}
\usage{
cytof_addToFCS_modified(data, rawFCSdir, origSampNames = NULL,
  analyzedFCSdir, transformed_cols = NULL, cluster_cols = NULL,
  clusterIDs = data, specifySampleNames = NULL, inLgclTrans = FALSE,
  newHeader = "cytofkit_")
}
\arguments{
\item{data}{The new data matrix to be added in.}

\item{rawFCSdir}{The directory containing the original fcs files.}

\item{origSampNames}{Vector of original names of samples, if samples were renamed.}

\item{analyzedFCSdir}{The directory to store the new fcs files.}

\item{transformed_cols}{The column name of the dimension transformed data in \code{data}.}

\item{cluster_cols}{The column name of the cluster data in \code{data}.}

\item{clusterIDs}{Table of cell cluster IDs for each clustering method}

\item{specifySampleNames}{Used only if sample names differ from those in raw fcs.}

\item{inLgclTrans}{If \verb{TRUE}, apply the inverse lgcl transformation to the the cluster data before saving}

\item{newHeader}{生成文件名前缀}
}
\value{
none
}
\description{
在已有FCS文件中增加额外的通道
}
\details{
The following functions are modified from cytofkit, please cite cytofik article if they were used in your pepline.
}
