% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_addToFCS_modified.R
\name{cytofAsinh}
\alias{cytofAsinh}
\title{cytofAsinh 定义数据转化函数}
\usage{
cytofAsinh(value, cofactor = 5)
}
\arguments{
\item{value}{一个vector，待转化的数据}

\item{cofactor}{一个数字做为辅助因子，vector中的数据要先除以cofactor，然后再进行Arcsin转化}
}
\description{
本函数来自cytofkit::cytofAsinh，用于Cytof数据的转化，实现去背景（减去1然后将负值随机化），和数据Arcsinh转化。
}
