% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{dif_seq_rainbow}
\alias{dif_seq_rainbow}
\title{dif_seq_rainbow}
\usage{
dif_seq_rainbow(n)
}
\arguments{
\item{n}{需要产生的颜色种类数量}
}
\value{
返回指定数量的颜色序列
}
\description{
生成一系列差异较大的颜色，适合于在散点图上显示不同population的细胞
}
