% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Abundance.R
\name{draw_abundance_boxplot}
\alias{draw_abundance_boxplot}
\title{画各个Cluster Abundance的boxplot}
\usage{
draw_abundance_boxplot(cluster_stat, cluster_id = NA)
}
\arguments{
\item{cluster_stat}{stat_by_cluster()函数生成的list数据}

\item{cluster_id}{指定输出cluster的id，例如：如果要输出前五个cluster，cluster_id=c(1:5); 如果要输出全部cluster，保持其默认值NULL即可。}
}
\value{
none
}
\description{
对cluster进行统计分析时，过少的细胞数有可能造成统计结果的不准确，本函数统计分析方法分析每个cluster的绝对数量是否符合要求
}
