% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Abundance.R
\name{draw_abundance_volcano}
\alias{draw_abundance_volcano}
\title{生成Cluster Aundance的火山图}
\usage{
draw_abundance_volcano(cluster_stat, cond1, cond2, xlimit = NULL,
  ylimit = NULL, dif.level = 2,
  output_dir = paste0("Aundance_volcano_plots_", cond1, "_vs_", cond2),
  stat.paired = NULL, stat.method = NULL, set.equal.var = NULL,
  conf.level = NULL, p.adjust.method = NULL)
}
\arguments{
\item{cluster_stat}{stat_by_cluster()函数生成的list数据}

\item{cond1}{major_cond的一个，火山图（Volcano map）靠左一侧的条件，例如"B_Adjacent"}

\item{cond2}{major_cond的一个，火山图（Volcano map）靠右一侧的条件，例如"A_Tumor"}

\item{xlimit}{一个包含两个数字的向量（vector），例如c(-5,+5),用于手动指定火山图横坐标范围，默认值为NULL，此时会自动设置xlimit}

\item{ylimit}{一个包含两个数字的向量（vector），例如c(0,3.5),用于手动指定火山图纵坐标范围，默认值为NULL，此时会自动设置xlimit}

\item{dif.level}{一个数字，手动设置两组marker差异的的阈值大小，例如取默认值2时，两组均值相差2倍以上算显著（简单的说就是设置火山图上的两条竖线位置）}

\item{output_dir}{输出数据文件夹名称

如使用全局统计默认参数，以下参数无需设置}

\item{stat.paired}{逻辑变量，TRUE或者FALSE，表示是否是成对检验}

\item{stat.method}{一个字符串，指定统计分析的方法，取值有三种："auto"，“t-test"，“wilcox-test”,当取值为"auto"时，即为测试模式，会对数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。}

\item{set.equal.var}{一个字符串，指定方差齐性，取值有三种："TRUE","FALSE"(默认),"auto",选择"auto"时即可以进行方差齐性测试；}

\item{conf.level}{显著性水平，默认0.95,（即p<0.05为显著），调整该数值可以调整火山图上的横线位置}

\item{p.adjust.method}{对p值进行校正的方法："BH"(默认，推荐) "holm", "hochberg", "hommel", "bonferroni","BY","fdr", "none"}
}
\value{
none
}
\description{
依据stat_by_cluster生成的list数据，生成火山图，图中颜色代表不同的Marker，数字和字母等符号代表不同的cluster；
结果采用了与SPADEvizR相同的形式，标注有“cite from SPADEVizR”或者“Adapted from SPADEVizR” ，使用时请注意在文章中引用原始文献。
}
