% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{draw_density_plots}
\alias{draw_density_plots}
\title{draw_density_plots}
\usage{
draw_density_plots(combined_data_plot, groups, all_markers,
  cluster_color = dif_seq_rainbow, cluster_name = "metacluster",
  output_dir = "cluster_density_plots", major_cond = "Tissue_Type",
  cluster_id = NULL, xlim = NULL, free_x = F)
}
\arguments{
\item{combined_data_plot}{数据框或者矩阵，附带有(meta)cluster、降维结果(t_sne_1,t_sne_2)的表达矩阵}

\item{groups}{实验组的Metadata信息,其中必需包含列名：”Short_name“}

\item{cluster_color}{tSNE图的亚群的配色方案，默认dif_seq_rainbow}

\item{cluster_name}{用过分析的cluster名称}

\item{output_dir}{输出数据文件夹名称}

\item{major_cond}{选择groups中一个列名做为展现差异的major_cond}

\item{cluster_id}{指定显示的cluster}

\item{xlim}{指定x轴范围，例如：c(0,5)，只在free_x=F的时候有效}

\item{free_x}{True或者FALSE，设置是否每张图X-轴取值范围是否自动，建议设置为F}
}
\description{
自动生成每个marker的直方图（密度图）；
}
