% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{draw_expr_volcano}
\alias{draw_expr_volcano}
\title{生成marker表达的火山图}
\usage{
draw_expr_volcano(
  cluster_stat,
  cond1,
  cond2,
  Usecolorset = dif_seq_rainbow,
  show_adjust_p = FALSE,
  cluster_id = NULL,
  xlimit = NULL,
  ylimit = NULL,
  dif.level = 2,
  output_dir = paste0("Expr_volcano_plots_", cond1, "_vs_", cond2),
  stat.paired = NULL,
  stat.method = NULL,
  set.equal.var = NULL,
  conf.level = NULL,
  p.adjust.method = NULL
)
}
\arguments{
\item{cluster_stat}{stat_by_cluster()函数生成的list数据}

\item{cond1}{major_cond的一个，火山图（Volcano map）靠左一侧的条件}

\item{cond2}{major_cond的一个，火山图（Volcano map）靠右一侧的条件}

\item{Usecolorset}{火山图采用的配色方案 默认dif_seq_rainbow, 其他取值：rainbow，primary.color或者其他函数}

\item{show_adjust_p}{一个逻辑变量， 取值TRUE时火山图显示校正后的p值，取值FALSE时显示原始p值}

\item{cluster_id}{指定输出cluster的id，例如：如果要输出前五个cluster，cluster_id=c(1:5); 如果要输出全部cluster，保持其默认值NULL即可。}

\item{xlimit}{一个包含两个数字的向量（vector），例如c(-5,+5),用于手动指定火山图横坐标范围，默认值为NULL，此时会自动设置xlimit}

\item{ylimit}{一个包含两个数字的向量（vector），例如c(0,3.5),用于手动指定火山图纵坐标范围，默认值为NULL，此时会自动设置xlimit}

\item{dif.level}{一个数字，手动设置两组marker差异的的阈值大小，例如取默认值2时，两组均值相差2倍以上算显著（简单的说就是设置火山图上的两条竖线位置）}

\item{output_dir}{输出数据文件夹名称
如使用全局统计默认参数，以下参数无需设置}

\item{stat.paired}{逻辑变量，TRUE或者FALSE，表示是否是成对检验}

\item{stat.method}{一个字符串，指定统计分析的方法，取值有三种："auto"，“t-test"，“wilcox-test”,当取值为"auto"时，即为测试模式，会对数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。}

\item{set.equal.var}{一个字符串，指定方差齐性，取值有三种："TRUE","FALSE"(默认),"auto",选择"auto"时即可以进行方差齐性测试；}

\item{conf.level}{显著性水平，默认0.95,（即p<0.05为显著），调整该数值可以调整火山图上的横线位置}

\item{p.adjust.method}{对p值进行校正的方法："BH"(默认，推荐) "holm", "hochberg", "hommel", "bonferroni","BY","fdr", "none"}
}
\value{
none
}
\description{
依据stat_by_cluster生成的list数据，生成火山图，图中颜色代表不同的Marker，数字和字母等符号代表不同的cluster；
}
