% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R, R/equal_sample.R
\name{draw_tsne_figs}
\alias{draw_tsne_figs}
\title{draw_tsne_figs}
\usage{
draw_tsne_figs(
  combined_data_plot,
  groups,
  cluster_color = dif_seq_rainbow,
  cluster_name = "metacluster",
  output_dir = "cluster_tsne_plots",
  major_cond = "Tissue_Type",
  reduction_dm1 = "tsne_1",
  reduction_dm2 = "tsne_2",
  dot_size = 4,
  text_size = 30,
  show_contour = FALSE,
  contour_line_size = 0.5,
  contour_bins = 25,
  contour_colour = "grey",
  group_seq = NULL,
  output_format = "pdf"
)

draw_tsne_figs(
  combined_data_plot,
  groups,
  cluster_color = dif_seq_rainbow,
  cluster_name = "metacluster",
  output_dir = "cluster_tsne_plots",
  major_cond = "Tissue_Type",
  reduction_dm1 = "tsne_1",
  reduction_dm2 = "tsne_2",
  dot_size = 4,
  text_size = 30,
  show_contour = FALSE,
  contour_line_size = 0.5,
  contour_bins = 25,
  contour_colour = "grey",
  group_seq = NULL,
  output_format = "pdf"
)
}
\arguments{
\item{combined_data_plot}{数据框或者矩阵，附带有(meta)cluster聚类、降维结果(t_sne_1,t_sne_2)的表达矩阵}

\item{groups}{实验组的Metadata信息,其中必需包含列名：”Short_name“}

\item{cluster_color}{tSNE图的亚群的配色方案，默认dif_seq_rainbow}

\item{cluster_name}{选择groups中一个列名做为展现差异的major_cond}

\item{output_dir}{输出数据文件夹名称}

\item{major_cond}{选择groups中一个列名做为展现差异的major_cond}

\item{reduction_dm1}{combined_data_plot降维产生的维度，默认"tsne_1"}

\item{reduction_dm2}{combined_data_plot降维产生的维度，默认"tsne_2"}

\item{dot_size}{散点图点的大小，默认是4}

\item{show_contour}{布尔型变量，决定是否显示等高线背景}

\item{contour_line_size}{设置等高线的宽度，默认0.5}

\item{contour_bins}{设置等高线的层数，默认25}

\item{contour_colour}{设置等高线的颜色，默认"grey"}

\item{group_seq}{设置各组顺序，格式实例： group_seq=c("PBMC","Biopsy")，注：括号内为各组名称；}

\item{output_format}{输出数据文件格式："tiff"和“pdf”,默认"pdf"}

\item{edge_data}{画网络图的时候，存储网络连接数据}

\item{cluster_barplot_data}{包含cluster通道信息的Vector，用来生成cluster的丰度直方图}

\item{cluster_labels_size}{控制cluster label的大小，默认30}

\item{show_axis}{是否显示坐标轴，默认是TRUE}

\item{show_cluster_labels}{决定是否显示cluster label，TRUE 显示Label，FALSE 不显示}

\item{groups_to_show}{指定要显示的condition（major_cond中的一个或者多个），如不设置直接统计全部；}

\item{edge_line_size}{连线的宽度}

\item{edge_layer}{连线的图层位置，"back"和"front"，决定连线图层在散点图的后面或前面}
}
\description{
自动生成系列tSNE图；

自动生成系列tSNE图；
}
