% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{draw_tsne_heatmaps}
\alias{draw_tsne_heatmaps}
\title{draw_tsne_heatmaps}
\usage{
draw_tsne_heatmaps(combined_data_plot = combined_data_plot,
  heatmap_tsne_markers = NULL, single_file = FALSE,
  output_dir = "tsne_heatmap", reduction_dm1 = "tsne_1",
  reduction_dm2 = "tsne_2")
}
\arguments{
\item{combined_data_plot}{数据框或者矩阵，附带有(meta)cluster、降维结果(t_sne_1,t_sne_2)的表达矩阵}

\item{heatmap_tsne_markers}{vector，指定marker的名称,默认NULL，此时将生成所有通道的heatmap}

\item{single_file}{是否把所有heatmap合成单个文件，默认FALSE}

\item{output_dir}{输出数据文件夹名称}

\item{reduction_dm1}{combined_data_plot降维产生的维度，默认"tsne_1"}

\item{reduction_dm2}{combined_data_plot降维产生的维度，默认"tsne_2"}
}
\description{
生成各个marker的tSNE；
}
