% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R, R/equal_sample.R
\name{equal_sample}
\alias{equal_sample}
\title{equal_sample}
\usage{
equal_sample(
  x,
  sample_type = "by_file",
  sample_method = "ceil",
  sample_num = 1000,
  cluster_name = "metaCluster",
  groups = NULL,
  sample_cond = NULL
)

equal_sample(
  x,
  sample_type = "by_file",
  sample_method = "ceil",
  sample_num = 1000,
  cluster_name = "metaCluster",
  groups = NULL,
  sample_cond = NULL
)
}
\arguments{
\item{x}{需要进行Sampleing表达数据矩阵,除了表达数据，还包括两个通道：“File_ID”,以及聚类产生的cluster通道。}

\item{sample_type}{Sample的方式，#sample_type="by_cluster",从每个cluster抽取相同细胞；sample_type="by_file"，从每个样本抽取相同细胞}

\item{sample_method}{默认"ceil",   #两种取值："ceil", "all"，取"all"时，采取所有细胞}

\item{sample_num}{每个样本或者每个cluster抽取的细胞数量；}

\item{cluster_name}{cluster通道的名称}
}
\value{
返回值一个矩阵，包含所有sample取得的表达数据

返回值一个矩阵，包含所有sample取得的表达数据
}
\description{
从各个Meta cluster或文件中抽取定量细胞进行可视化
可视化前的Sampling有两个功能，第一，减少后续用于降维分析的细胞总数，一般推荐在10万以内，以保证降维效果。
第二， 解决原始数据中文件或者cluster存在细胞数目不均衡的情况。增加对小文件或cluster的可见性。

从各个Meta cluster或文件中抽取定量细胞进行可视化
可视化前的Sampling有两个功能，第一，减少后续用于降维分析的细胞总数，一般推荐在10万以内，以保证降维效果。
第二， 解决原始数据中文件或者cluster存在细胞数目不均衡的情况。增加对小文件或cluster的可见性。
}
