% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{ggtable}
\alias{ggtable}
\title{简易表格}
\usage{
ggtable(data, show_row_name = TRUE, show_col_name = TRUE)
}
\arguments{
\item{data}{待输入的dataframe}

\item{show_row_name}{TRUE/FALSE, 决定是否在输出的表格中显示行名}

\item{show_col_name}{TRUE/FALSE, 决定是否在输出的表格中显示列名}
}
\value{
返回一个ggplot对象
}
\description{
：
将输入dataframe的内容化成表格
}
