% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_treatment.R
\name{groups2csv}
\alias{groups2csv}
\title{groups2csv    生成所有文件的列表}
\usage{
groups2csv(filename, add_col = c("Short_name"))
}
\arguments{
\item{markernames}{combind_data_raw的名称}

\item{add_cols}{要在all_markers.csv中加入的列名}
}
\description{
groups2csv    生成所有文件的列表
}
