% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_treatment.R
\name{markers2csv}
\alias{markers2csv}
\title{markers2csv  生成所有marker的名称列表}
\usage{
markers2csv(
  rawname,
  markername,
  add_col = c("transform", "tSNE", "PhenoGraph", "tsne_heatmap", "expr_para",
    "heatmap")
)
}
\arguments{
\item{rawname}{fcs文件中原始名称}

\item{markernames}{combind_data_raw的名称}

\item{add_cols}{要在all_markers.csv中加入的列名}
}
\description{
markers2csv  生成所有marker的名称列表
}
