% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{metacluster_multi}
\alias{metacluster_multi}
\title{metacluster_multi}
\usage{
metacluster_multi(data, method = "metaClustering_hclust",
  nCluster = NULL, max = 50, k = 15)
}
\arguments{
\item{data}{包含所有cluster的表达矩阵，dataframe或者矩阵格式}

\item{method}{要使用的聚类方法名称，取值有“PhenoGraph”，“metaClustering_consensus”,“metaClustering_hclust”, “metaClustering_kmeans”}

\item{nCluster}{使用“metaClustering_consensus”和“metaClustering_kmeans”两种方法时，用其指定产生metaCluster的数量}

\item{max}{使用“metaClustering_hclust”时，确定一个metacluster的上限，函数将在此下寻找合适的metacluster数量}

\item{k}{使用“PhenoGraph"时，设置knn网络的k值，默认值为15}
}
\value{
一个数组（Numeric array）,包含对应每个cluster的metacluster数值
}
\description{
利用FlowSOM或其他方法进行聚类以后，可以对产生的cluster进一步聚类，这些新产生的类被称为“metacluster”。
metacluster_multi可以利用FlowSOM自带的metacluster函数，以及PhenoGraph算法产生metacluster。
#'
}
