% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R, R/equal_sample.R
\name{One_SENSE_report}
\alias{One_SENSE_report}
\title{One_SENSE_report}
\usage{
One_SENSE_report(
  combined_data_sampled,
  perplexity = 30,
  max_iter = 1000,
  all_markers = all_markers,
  output_dir = "cluster_onesense_plots"
)

One_SENSE_report(
  combined_data_sampled,
  perplexity = 30,
  max_iter = 1000,
  all_markers = all_markers,
  output_dir = "cluster_onesense_plots"
)
}
\arguments{
\item{combined_data_sampled}{数据框或者矩阵，待处理的表达数据}

\item{perplexity}{tsne降维参数，困惑度}

\item{max_iter}{tsne降维参数，迭代次数}

\item{all_markers}{带有One_SENSE_F和One_SENSE_P的通道选择marker}

\item{output_dir}{输出的目录名称}
}
\description{
进行OneSense分析，并进行画出两个坐标轴方向的Heatmap；

进行OneSense分析，并进行画出两个坐标轴方向的Heatmap；
}
