% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{PG_elbow}
\alias{PG_elbow}
\title{Elbow Test}
\usage{
PG_elbow(x, k_from = 5, k_to = 100, k_step = 5)
}
\arguments{
\item{x}{需要进行测试的表达数据矩阵}

\item{k_from}{自然数，测试中k的取值范围的起始点}

\item{k_to}{自然数，测试中k的取值范围的结束点}

\item{k_step}{自然数，k取值的步长}
}
\value{
返回值一个矩阵，记录每个k值对应的cluster数目
}
\description{
Phenograph 的k值是指在计算k 近邻网络(knn)的所取得k值，它的意义是网络中每个细胞的最近邻居数。K值越大，最后得到的cluster越少，反之越大。
因此过大的k值会引起“underclustering”，导致cluster偏少，过小的k值则会导致“overclustering”,导致cluster偏高。
选择合适的k值就是在两者之间寻求平衡。比较直观的方法就是在k-cluster图中，找到曲线的 “elbow piont”。
}
