% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{remove_extremum}
\alias{remove_extremum}
\title{去除极值}
\usage{
remove_extremum(value)
}
\arguments{
\item{value}{待处理的数据}
}
\value{
a vector of numeric values containing treated data
}
\description{
：
去除极值，将大于0.95分位数的数值赋值为0.95分位数，小于0.05分为数的数值赋值为0.05分位数
}
