% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{simpleAsinh}
\alias{simpleAsinh}
\title{Arcsinh 转化函数}
\usage{
simpleAsinh(value, cofactor = 5)
}
\arguments{
\item{value}{the data need to be transformed,a vector of numeric values.}

\item{cofactor}{Cofactor for asinh transformation, default 5 for mass cytometry data.}
}
\value{
a vector of numeric values containing transformed data
}
\description{
：
这是一个数据转化函数，由Cytofkit::cytofAsinh() 函数修改而来，根据需要进行了简化，主要去掉了原函数关于扣除背景和随机化的部分。
}
