% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Expr.R
\name{stat_by_cluster}
\alias{stat_by_cluster}
\title{Statistics of clusters(Abundance and Experssion)}
\usage{
stat_by_cluster(combined_data_raw, all_markers,
  cluster_name = "metacluster", summerise_method = "median", groups,
  major_cond, heatmap_ctrl, stat.paired = TRUE, stat.method = "t-test",
  set.equal.var = "FALSE", conf.level = 0.95, p.adjust.method = "BH")
}
\arguments{
\item{combined_data_raw}{在Pepline前面部分读入并合并的表达数据}

\item{all_markers}{从 “all_markers.csv”读入的dataframe，包含marker的选择信息}

\item{cluster_name}{统计所依据的聚类参数名称，例如"PhenoGraph"、"metacluster"}

\item{summerise_method}{统计过程中使用的合计方法：有"mean"或者 "median"两种}

\item{groups}{groups 从“all_sample.csv”中读入的dataframe，包含实验分组信息（也就是元数据）

Heatmap 参数}

\item{major_cond}{在groups所提供的各种分组中，选择一种做为统计差异的依据，例如"Tissue_Type"}

\item{heatmap_ctrl}{在major_cond中选择一个在Heatmap中显示做为control的组名，例如"A_Tumor"
下面5个参数是为后面统计分析中设置的默认值}

\item{stat.paired}{逻辑变量，TRUE或者FALSE，表示是否是成对检验}

\item{stat.method}{一个字符串，指定统计分析的方法，取值有三种："auto"，“t-test"，“wilcox-test”,当取值为"auto"时，即为测试模式，会对数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。}

\item{set.equal.var}{一个字符串，指定方差齐性，取值有三种："TRUE","FALSE"(默认),"auto",选择"auto"时即可以进行方差齐性测试；}

\item{conf.level}{显著性水平，默认0.95,（即p<0.05为显著）}

\item{p.adjust.method}{对p值进行校正的方法："BH"(默认，推荐) "holm", "hochberg", "hommel", "bonferroni","BY","fdr", "none"}
}
\value{
a List containing the data which needed in drawing of heatmap, boxplot and volcano chart
}
\description{
统计不同分组中各个cluster中marker表达的差异
}
