% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_cytof_addToFCS_modified.R
\name{cytof_addToFCS_modified}
\alias{cytof_addToFCS_modified}
\title{cytof_addToFCS_modified}
\usage{
cytof_addToFCS_modified(
  data,
  rawFCSdir,
  origSampNames = NULL,
  analyzedFCSdir,
  newHeader = ""
)
}
\arguments{
\item{data}{一个datafram，包含想要写入FCS文件的各个参数}

\item{rawFCSdir}{原始FCS文件所在的目录}

\item{origSampNames}{原始的FCS文件名称}

\item{analyzedFCSdir}{生成FCS文件存放的目录名}

\item{newHeader}{一个字符串，做为新生成文件名的前缀，方便文件识别}
}
\value{
返回指定数量的颜色序列
}
\description{
本函数修改自cytofkit::cytof_addToFCS,根据数据分析的需要对原函数进行了简化和修改。功能是将降维和聚类分析的数据整合到FCS文件中。
请注意引用Cytofkit的原始文献，详情运行 citation("cytofkit") 查看。
}
\details{
Notice: The following functions are modified or adapted from cytofkit, please cite original articles if they were used in your pepline.
For details, please enter citation("cytofkit") in console lines.
}
