% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Abundance.R
\name{draw_abundance_boxplot}
\alias{draw_abundance_boxplot}
\title{画各个Cluster Abundance的boxplot}
\usage{
draw_abundance_boxplot(cluster_stat, cluster_id = NULL,
  show_pvalues = TRUE, use_shiny_para = FALSE, outputfig = TRUE,
  colorset = brewer_color_sets, color_cond = NULL, boxplot_ctrl,
  comparisons = NULL, output_dir = paste0("Abundance_boxplot"),
  group_seq = NULL, stat.paired = NULL, stat.method = NULL,
  set.equal.var = NULL, conf.level = NULL, p.adjust.method = NULL)
}
\arguments{
\item{cluster_stat}{stat_by_cluster()函数生成的list数据}

\item{cluster_id}{指定输出cluster的id，例如：如果要输出前五个cluster，cluster_id=c(1:5); 如果要输出全部cluster，保持其默认值NULL即可。}

\item{show_pvalues}{逻辑变量，TRUE或者FALSE，决定是否在图上显示p Value}

\item{colorset}{boxplot采用的配色方案：默认 brewer_color_sets, 也可以选择其他公开生成色阶的函数例如 rainbow等}

\item{color_cond}{指定依据哪个条件进行着色}

\item{boxplot_ctrl}{指定boxplot的control组名称，例如“PBMC”}

\item{comparisons}{list变量，用来指定需要显示p值的组别，例如list(c("PBMC","Biopsy")，c("PBMC","Tumor"))就是要分别显示PBMC与Biopsy和Tumor两组之间的p值；如果comparisons=NULL，则显示各组与对照组的p值}

\item{output_dir}{输出数据文件夹名称，默认"Abundance_boxplot"
如使用统计默认参数，以下参数无需设置}

\item{group_seq}{设置各组顺序，格式实例： group_seq=c("PBMC","Biopsy")，注：括号内为各组名称；如不设置，默认与总体参数保持一致}

\item{stat.paired}{逻辑变量，TRUE或者FALSE，表示是否是成对检验}

\item{stat.method}{一个字符串，指定统计分析的方法，取值有三种："auto"，“t-test"，“wilcox-test”,当取值为"auto"时，即为测试模式，会对数据进行正态分布和方差齐性测试，自动选择适合的统计方法，为正式分析过程中的统计方法选择提供依据。}

\item{set.equal.var}{一个字符串，指定方差齐性，取值有三种："TRUE","FALSE"(默认),"auto",选择"auto"时即可以进行方差齐性测试；}

\item{conf.level}{显著性水平，默认0.95,（即p<0.05为显著）}

\item{p.adjust.method}{对p值进行校正的方法："BH"(默认，推荐) "holm", "hochberg", "hommel", "bonferroni","BY","fdr", "none"}

\item{colorset}{boxplot采用的配色方案：默认 brewer_color_sets, 也可以选择其他公开生成色阶的函数例如 rainbow等}
}
\value{
a list containing ggplot2 objects
}
\description{
对cluster进行统计分析时，过少的细胞数有可能造成统计结果的不准确，本函数统计分析方法分析每个cluster的绝对数量是否符合要求
}
