% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R
\name{draw_tsne_figs}
\alias{draw_tsne_figs}
\title{draw_tsne_maps}
\usage{
draw_tsne_figs(combined_data_plot, groups,
  cluster_color = dif_seq_rainbow, cluster_name = "metacluster",
  output_dir = "cluster_tsne_plots", major_cond = "Tissue_Type",
  reduction_dm1 = "tsne_1", reduction_dm2 = "tsne_2",
  group_seq = NULL)
}
\arguments{
\item{combined_data_plot}{数据框或者矩阵，附带有(meta)cluster聚类、降维结果(t_sne_1,t_sne_2)的表达矩阵}

\item{groups}{实验组的Metadata信息,其中必需包含列名：”Short_name“}

\item{cluster_color}{tSNE图的亚群的配色方案，默认dif_seq_rainbow}

\item{cluster_name}{选择groups中一个列名做为展现差异的major_cond}

\item{output_dir}{输出数据文件夹名称}

\item{reduction_dm1}{combined_data_plot降维产生的维度，默认"tsne_1"}

\item{reduction_dm2}{combined_data_plot降维产生的维度，默认"tsne_2"}

\item{group_seq}{设置各组顺序，格式实例： group_seq=c("PBMC","Biopsy")，注：括号内为各组名称；}
}
\description{
自动生成系列tSNE图；
}
