% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM_tSNE_Output.R, R/equal_sample.R
\name{draw_tsne_heatmaps}
\alias{draw_tsne_heatmaps}
\title{draw_tsne_heatmaps}
\usage{
draw_tsne_heatmaps(
  combined_data_plot,
  heatmap_tsne_markers = NULL,
  trans_method = "simpleAsinh",
  single_file = FALSE,
  groups = groups,
  major_cond = NULL,
  dot_size = 4,
  dot_colours = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F",
    "yellow", "#FF7F00", "red", "#7F0000")),
  show_legend = TRUE,
  show_scale = TRUE,
  output_dir = "tsne_heatmap",
  output_format = "tiff",
  reduction_dm1 = "tsne_1",
  reduction_dm2 = "tsne_2"
)

draw_tsne_heatmaps(
  combined_data_plot,
  heatmap_tsne_markers = NULL,
  trans_method = "simpleAsinh",
  single_file = FALSE,
  groups = groups,
  major_cond = NULL,
  dot_size = 4,
  dot_colours = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F",
    "yellow", "#FF7F00", "red", "#7F0000")),
  show_legend = TRUE,
  show_scale = TRUE,
  output_dir = "tsne_heatmap",
  output_format = "tiff",
  reduction_dm1 = "tsne_1",
  reduction_dm2 = "tsne_2"
)
}
\arguments{
\item{combined_data_plot}{数据框或者矩阵，附带有(meta)cluster、降维结果(t_sne_1,t_sne_2)的表达矩阵}

\item{heatmap_tsne_markers}{vector，指定marker的名称,默认NULL，此时将生成所有通道的heatmap}

\item{trans_method}{数据转化方式，有四种："CytofAsinh"，"simpleAsinh"；"0_to_Max"，所有Marker的信号强度除以最大值，线性转换，通过除以各通道信号的最大值把数值scale到0~1；"Min_to_Max"，线性转换，最小值转换成0，最大值转换成1，最大限度展现population的表达差异}

\item{single_file}{是否把所有heatmap合成单个文件，默认FALSE}

\item{groups}{实验组的Metadata信息,其中必需包含列名：”Short_name“}

\item{major_cond}{选择groups中一个列名做为展现差异的major_cond}

\item{dot_size}{图中点的大小，默认是4}

\item{dot_colours}{设置colorbar的颜色，默认为jet color函数}

\item{show_legend}{是否显示legend（colorbar）}

\item{show_scale}{是否显示X 轴和Y轴的刻度，默认是TRUE}

\item{output_dir}{输出数据文件夹名称，默认是TRUE}

\item{output_format}{输出数据文件格式："tiff"和“pdf”}

\item{reduction_dm1}{combined_data_plot降维产生的维度，默认"tsne_1"}

\item{reduction_dm2}{combined_data_plot降维产生的维度，默认"tsne_2"}

\item{edge_data}{画网络图的时候，存储网络连接数据}

\item{groups_to_show}{vector，指定在图上显示哪些组的数据（组名应属于all_samples的major_cond列）,默认为NULL，显示所有组}

\item{legend_type}{legend(colorbar)是局部（"local"(默认)）,所有marker用不同的colorbar）还是全局（"global",所有marker共用一个colorbar）}

\item{legend_limit}{仅在legend_type是"global"的情况下有效，手动设置colorbar的上下限，取值为一个vector，例如c(0,5) 0为最小值，5为最大值}

\item{show_axis}{是否显示坐标轴，默认是TRUE}
}
\value{
none
}
\description{
生成各个marker的tSNE；

生成各个marker的tSNE；
}
